+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Removed
 - Creation of automatic PREMIS events when adding imported metadata to structural map. This functionality is moved over to ``siptool-ng``.
 - ``from_directory_structure`` method from ``StructureMap`` class. This functionality is moved over to ``siptool-ng``.
 - Rename MetadataBase class to Metadata
+- Make Metadata an abstract base class
 
 0.3.0 - 2024-03-27
 ------------------